範囲攻撃の追加
GameObjectにis_AoEを追加
このフラグをtrueにすると範囲攻撃になる
逆にfalseにすると単体攻撃になる

+++ ProjectTRM/Objects/Character/Enemy/EnemyBase.h
@@ -12,7 +12,7 @@ protected:
 	float recovery_time;	// 硬直時間
 	float recovery_flame;	// 硬直時間の計算
 	float speed;			// 歩く速さ
-	int Damage;				// ダメージ
+	float Damage;				// ダメージ
 	bool old_light;			// 前回のライトフラグ
 
 	// エフェクト
